+++ truffle/src/com.oracle.truffle.polyglot/src/com/oracle/truffle/polyglot/PolyglotImpl.java
@@ -437,5 +437,4 @@ public final class PolyglotImpl extends AbstractPolyglotImpl {
         }
 
     }
-
 }